backend/request_access/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-auth==2.*
flask>=3.0
